requirements.txt
# UI
streamlit==1.44.1

# OCR
pytesseract==0.3.13
Pillow==11.2.1
pymupdf==1.27.2.2
gradio_client>=1.5.0

# Image processing
opencv-python-headless>=4.9.0
numpy>=1.24.0

# AI
anthropic==0.52.0
groq>=0.11.0

# PDF Export
fpdf2==2.8.3

# Environment
python-dotenv==1.1.0

# Analytics
plotly>=5.0.0

# Data
# sqlite3 is built-in to Python
